test: full test coverage (#53)
* test: update mxnet

* test: launch tasks entirely in python

* test: auto clean temp files

* test: update

* test: update

* test: update

* test: update

* test: update

* test: update

* test: update

* test: fix natural dithering

* test: update

* test: update

+++ byteps/common/compressor/utils.h
@@ -87,7 +87,7 @@ class XorShift128PlusBitShifterRNG {
   double Rand() { return double(xorshift128p()) / MAX; }
 
   // Bernoulli Distributation
-  bool Bernoulli(double p) { return xorshift128p() < uint64_t(p * MAX); }
+  bool Bernoulli(double p) { return xorshift128p() < p * MAX; }
 
   void set_seed(uint64_t seed) { _state = {seed, seed}; }
